refactor wall fragment
