+++ app/build.gradle
@@ -44,24 +44,24 @@ dependencies {
     compile "com.google.firebase:firebase-storage:${firebaseVersion}"
     compile "com.google.firebase:firebase-auth:${firebaseVersion}"
 
-    compile 'com.google.android.gms:play-services-auth:9.4.0'
 
     apt "com.jakewharton:butterknife-compiler:${butterknifeVersion}"
     compile "com.jakewharton:butterknife:${butterknifeVersion}"
 
+
+
+
+    apt 'org.parceler:parceler:1.1.1'
+    compile 'com.google.android.gms:play-services-auth:9.4.0'
     compile 'com.jakewharton.timber:timber:3.1.0'
-
     compile 'com.astuetz:pagerslidingtabstrip:1.0.1'
-
     compile 'com.github.bumptech.glide:glide:3.7.0'
     compile 'jp.wasabeef:glide-transformations:2.0.1'
-
     compile 'de.hdodenhof:circleimageview:2.1.0'
     compile 'org.parceler:parceler-api:1.1.1'
-    apt 'org.parceler:parceler:1.1.1'
 
     compile 'com.github.jkwiecien:EasyImage:1.2.3'
-
+    compile 'com.android.support:support-v4:24.2.0'
 }
 
 apply plugin: 'com.google.gms.google-services'
