+++ app/src/main/res/values/strings.xml
@@ -13,4 +13,7 @@
     <string name="firebase_tasks_table">Tasks</string>
     <string name="task_in_bid_process">IN_BID_PROCESS</string>
     <string name="accepted_offer_id">acceptedOfferId</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
